Refactored some methods in LoggedInView and removed psvm
